Create CONTRIBUTING.md
A CONTRIBUTING file to help developers understand how they can contribute to the project. In this file there are branches where we commonly write the new features of the library or those where we fix the errors instead of directly working on the main branch, also the file contains a short coding style guide to keep as much as possible the same style in the project.